+++ lib/pronto.rb
@@ -3,6 +3,8 @@ require 'octokit'
 require 'gitlab'
 require 'forwardable'
 
+require 'pronto/gem_names'
+
 require 'pronto/config_file'
 require 'pronto/config'
 
@@ -42,18 +44,5 @@ module Pronto
     puts formatted if formatted
 
     result
-  end
-
-  def self.gem_names
-    gems = Gem::Specification.find_all.select do |gem|
-      if gem.name =~ /^pronto-/
-        true
-      elsif gem.name != 'pronto'
-        runner_path = File.join(gem.full_gem_path, "lib/pronto/#{gem.name}.rb")
-        File.exist?(runner_path)
-      end
-    end
-
-    gems.map { |gem| gem.name.sub(/^pronto-/, '') }.uniq.sort
   end
 end