+++ lib/pronto/rake_task/travis_pull_request.rb
@@ -37,7 +37,7 @@ module Pronto
         pull_request = client.pull_request(repo_slug, pull_id)
         formatter = ::Pronto::Formatter::GithubFormatter.new
 
-        ::Pronto.gem_names.each { |gem_name| require "pronto/#{gem_name}" }
+        ::Pronto::GemNames.new.to_a.each { |gem_name| require "pronto/#{gem_name}" }
         ::Pronto.run(pull_request.base.sha, '.', formatter)
       end